+++ src/EndPoint/Token/RevocationEndPoint.php
@@ -2,6 +2,7 @@
 
 namespace Parroauth2\Client\EndPoint\Token;
 
+use Parroauth2\Client\Authentication\ClientAuthenticationMethodInterface;
 use Parroauth2\Client\ClientInterface;
 use Parroauth2\Client\EndPoint\CallableEndPointInterface;
 use Parroauth2\Client\EndPoint\EndPointParametersTrait;
@@ -124,13 +125,12 @@ class RevocationEndPoint implements CallableEndPointInterface
      */
     public function call(): ResponseInterface
     {
-        $request = $this->client->endPoints()
-            ->request('POST', $this)
-            ->withHeader(
-                'Authorization',
-                'Basic ' . base64_encode($this->client->clientId() . ':' . $this->client->secret())
-            )
-        ;
+        $client = $this->client;
+        $endPoints = $client->endPoints();
+        $request = $endPoints->request('POST', $this);
+        $authenticationMethod = $endPoints->authenticationMethod($this->name(), $client->option(ClientAuthenticationMethodInterface::OPTION_PREFERRED_METHOD));
+
+        $request = $authenticationMethod->apply($client, $request);
 
         $response = $this->client->provider()->sendRequest($request);
         $this->callResponseListeners($response);